fix: check npm dependencies to work as script
